merge: origin/main, and regenerate the chart engine
No source conflicts. The commit that matters is the REGENERATION: CI failed
this branch's chart-engine drift lock while a fresh generation on the branch
alone was byte-identical, which is the merge-ref trap — CI builds base + head,
and main's own committed engine is currently stale against its generator
(`Double(plot.h) / Double(n)` where a fresh run gives `plot.h / Double(n)`,
two Swift sites). The regen here corrects it on the merged tree.

#3404 fixes the same two lines on main directly, so whichever of the two lands
second will show an empty or trivially-conflicting diff. That is expected.

+++ packages/native/runtime-kotlin/src/main/kotlin/com/pyreon/runtime/PyreonChartEngine.kt
@@ -1271,7 +1271,7 @@ fun layoutGroupedBarsH(seriesValues: List<List<Double>>, plot: PyreonChartRect, 
         val raw = (seriesValues[s][i] ?: 0.0)
         val v = if (raw == raw) raw else zero
         val vx = scaleLinear(vDomain, plot.x, plot.x + plot.w, v)
-        out.add(StackSegment(rect = PyreonChartRect(x = if (vx < zeroX) vx else zeroX, y = gy + barH * s, w = (Math.abs(vx - zeroX)).toDouble(), h = barH), seriesIndex = s, datumIndex = i, value = v))
+        out.add(StackSegment(rect = PyreonChartRect(x = if (vx < zeroX) vx else zeroX, y = gy + barH * s, w = Math.abs(vx - zeroX), h = barH), seriesIndex = s, datumIndex = i, value = v))
       }
     }
     return out
